Change return type of several methods in machine-model from int to long
The changes made are a first try to tackle the problems as described in issue #426 on gitlab

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/size-type-refactoring@29229 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/cfa/simplification/ExpressionSimplificationVisitor.java
@@ -189,7 +189,7 @@ public class ExpressionSimplificationVisitor
 
     switch (idOperator) {
     case SIZEOF:
-      int size = machineModel.getSizeof(innerType);
+        long size = machineModel.getSizeof(innerType);
       return new CIntegerLiteralExpression(expr.getFileLocation(),
               expr.getExpressionType(), BigInteger.valueOf(size));
