+++ src/org/sosy_lab/cpachecker/util/predicates/precisionConverter/SymbolEncoding.java
@@ -150,7 +150,8 @@ public class SymbolEncoding {
     if (cType instanceof CSimpleType && ((CSimpleType)cType).getType().isFloatingPointType()) {
       fType = FormulaType.RationalType;
     } else {
-      int length = machineModel.getSizeof(cType) * machineModel.getSizeofCharInBits();
+      int length =
+          Math.toIntExact(machineModel.getSizeof(cType) * machineModel.getSizeofCharInBits());
       fType = FormulaType.getBitvectorTypeWithSize(length);
     }
     Type<FormulaType<?>> type = new Type<>(fType);
